refactor(web-app): move zod validateBySchema function to utils
