+++ packages/cube-frontend-web-app/src/pages/settings/_components/SlackChannels/useSlackChannelRowsErrorMap.ts
@@ -1,6 +1,6 @@
 import { SlackChannelPostRequest } from '@cube-frontend/api'
 import { useMemo } from 'react'
-import { ErrorRecord, validateBySchema } from '../validateBySchema'
+import { ErrorRecord, validateBySchema } from '@cube-frontend/web-app/utils/zod'
 import { SlackChannelRow, slackChannelSchema } from './slackChannelsUtils'
 
 export type SlackChannelRowError = ErrorRecord<SlackChannelPostRequest>
