+++ packages/cube-frontend-web-app/src/pages/settings/_components/email/recipients/useEmailRecipientRowsErrorMap.ts
@@ -1,5 +1,5 @@
 import { useMemo } from 'react'
-import { ErrorRecord, validateBySchema } from '../../validateBySchema'
+import { ErrorRecord, validateBySchema } from '@cube-frontend/web-app/utils/zod'
 import {
   EmailRecipientResponseWithoutStatus,
   EmailRecipientRow,
